+++ Sunlit_Meadow_Prototype00/Sunlit_Meadow_Prototype00/Block.h
@@ -39,7 +39,7 @@ public:
 
     void generateMeshFromModel(
         std::vector<WorldVertex>& vertices,
-        std::vector<Uint16>&   indices,
+        std::vector<Uint32>&   indices,
         int x, int y, int z
     );
